Joshuabg/crypto feature

+++ src/test/resources/createDB.sql
@@ -37,7 +37,7 @@ CREATE TABLE TransferHistory (
 CREATE TABLE CryptoHoldings (
   CustomerID varchar(255),
   CryptoName varchar(255),
-  CryptoAmount float
+  CryptoAmount decimal(30,18)
 );
 
 CREATE TABLE CryptoHistory (
@@ -45,5 +45,5 @@ CREATE TABLE CryptoHistory (
   Timestamp DATETIME,
   Action varchar(255) CHECK (Action IN ('Buy', 'Sell')),
   CryptoName varchar(255),
-  CryptoAmount float
+  CryptoAmount decimal(30,18)
 );
